runtime/ml_ops/start_finetuning.py
import boto3
import sagemaker
from sagemaker.estimator import Estimator
import os
import datetime

# Placeholder configuration - in real usage these would come from CDK outputs or ENV
# For now, we assume user fills these or they are passed in
ROLE_ARN = os.environ.get("SAGEMAKER_ROLE_ARN", "arn:aws:iam::ACCOUNT_ID:role/service-role/AmazonSageMaker-ExecutionRole-20200101T000001")
BUCKET_NAME = os.environ.get("S3_BUCKET_NAME", "my-sagemaker-bucket")
BASE_MODEL_ID = "meta-textgeneration-llama-3-8b" # SageMaker JumpStart ID equivalent

def start_finetuning():
    print(f"Starting Fine-Tuning Job for {BASE_MODEL_ID}...")
    
    # Initialize SageMaker Session
    sess = sagemaker.Session()
    
    # Define Training Input (Mock path)
    training_data_uri = f"s3://{BUCKET_NAME}/train/dataset.jsonl"
    
    # Define Hyperparameters
    hyperparameters = {
        "epoch": "1",
        "learning_rate": "0.0002",
        "instruction_tuned": "True",
        "chat_dataset": "True"
    }
    
    # NOTE: Actual Fine-tuning on Bedrock vs SageMaker JumpStart differs.
    # This script assumes SageMaker JumpStart generic approach or Custom container.
    # For Bedrock Fine-tuning, we would use boto3 bedrock.create_model_customization_job
    
    # Switching to Bedrock Fine-tuning implementation as it fits the context better
    bedrock = boto3.client('bedrock')
    
    job_name = f"finance-finetune-{datetime.datetime.now().strftime('%Y%m%d-%H%M%S')}"
    
    try:
        response = bedrock.create_model_customization_job(
            jobName=job_name,
            customModelName=f"finance-v1-{datetime.datetime.now().strftime('%Y%m%d')}",
            roleArn=ROLE_ARN,
            baseModelIdentifier="arn:aws:bedrock:us-east-1::foundation-model/anthropic.claude-3-sonnet-20240229-v1:0", # Example
            trainingDataConfig={
                "s3Uri": training_data_uri
            },
            outputDataConfig={
                "s3Uri": f"s3://{BUCKET_NAME}/output/"
            },
            hyperParameters={
                "epochCount": "1",
                "batchSize": "1",
                "learningRate": "0.0001"
            }
        )
        print(f"Job started: {response['jobArn']}")
        return response['jobArn']
        
    except Exception as e:
        print(f"Error creating job (expected if bucket/role invalid): {e}")

if __name__ == "__main__":
    start_finetuning()
